Using rake to facilitate simple test runs
This allows for a well known, and common default, use of rake.  In this
case, it will run the tests.

There is improvements that should be made to the rake file itself, but
this is enough to show how to use it.

This commit adds minitest and rake the the dependency list, but should
really separate between what is required for "development" and what is
required for "production"